fix all these tests
