+++ pallets/subtensor/tests/block_step.rs
@@ -14,11 +14,7 @@ fn test_loaded_emission() {
         let emission: Vec<u64> = vec![1000000000];
         add_network(netuid, tempo, 0);
         SubtensorModule::set_max_allowed_uids(netuid, n);
-        assert_ok!(SubtensorModule::do_set_emission_values(
-            <<Test as Config>::RuntimeOrigin>::root(),
-            netuids,
-            emission
-        ));
+        SubtensorModule::set_emission_values( &netuids, emission);
         for i in 0..n {
             SubtensorModule::append_neuron(netuid, &U256::from(i), 0);
         }
@@ -163,9 +159,9 @@ fn test_blocks_until_epoch() {
     });
 }
 
-/********************************************
-    block_step::adjust_registration_terms_for_networks tests
-*********************************************/
+// /********************************************
+//     block_step::adjust_registration_terms_for_networks tests
+// *********************************************/
 
 #[test]
 fn test_burn_adjustment() {
@@ -175,13 +171,14 @@ fn test_burn_adjustment() {
         let burn_cost: u64 = 1000;
         let adjustment_interval = 1;
         let target_registrations_per_interval = 1;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_adjustment_interval(netuid, adjustment_interval);
         SubtensorModule::set_target_registrations_per_interval(
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
+        assert_eq!( SubtensorModule::get_adjustment_interval(netuid), adjustment_interval ); // Sanity check the adjustment interval.
 
         // Register key 1.
         let hotkey_account_id_1 = U256::from(1);
@@ -220,6 +217,7 @@ fn test_burn_adjustment_with_moving_average() {
         let burn_cost: u64 = 1000;
         let adjustment_interval = 1;
         let target_registrations_per_interval = 1;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_adjustment_interval(netuid, adjustment_interval);
         SubtensorModule::set_target_registrations_per_interval(
@@ -227,7 +225,6 @@ fn test_burn_adjustment_with_moving_average() {
             target_registrations_per_interval,
         );
         // Set alpha here.
-        add_network(netuid, tempo, 0);
         SubtensorModule::set_adjustment_alpha(netuid, u64::MAX / 2);
 
         // Register key 1.
@@ -275,6 +272,7 @@ fn test_burn_adjustment_case_a() {
         let target_registrations_per_interval = 1;
         let start_diff: u64 = 10_000;
         let mut curr_block_num = 0;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_difficulty(netuid, start_diff);
         SubtensorModule::set_adjustment_interval(netuid, adjustment_interval);
@@ -282,7 +280,6 @@ fn test_burn_adjustment_case_a() {
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
 
         // Register key 1. This is a burn registration.
         let hotkey_account_id_1 = U256::from(1);
@@ -366,6 +363,7 @@ fn test_burn_adjustment_case_b() {
         let target_registrations_per_interval = 1;
         let start_diff: u64 = 10_000;
         let mut curr_block_num = 0;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_difficulty(netuid, start_diff);
         SubtensorModule::set_adjustment_interval(netuid, adjustment_interval);
@@ -373,7 +371,6 @@ fn test_burn_adjustment_case_b() {
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
 
         // Register key 1.
         let hotkey_account_id_1 = U256::from(1);
@@ -447,6 +444,7 @@ fn test_burn_adjustment_case_c() {
         let target_registrations_per_interval = 4; // Needs registrations < 4 to trigger
         let start_diff: u64 = 10_000;
         let mut curr_block_num = 0;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_difficulty(netuid, start_diff);
         SubtensorModule::set_adjustment_interval(netuid, adjustment_interval);
@@ -454,7 +452,6 @@ fn test_burn_adjustment_case_c() {
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
 
         // Register key 1. This is a BURN registration
         let hotkey_account_id_1 = U256::from(1);
@@ -538,6 +535,7 @@ fn test_burn_adjustment_case_d() {
         let target_registrations_per_interval = 4; // Needs registrations < 4 to trigger
         let start_diff: u64 = 10_000;
         let mut curr_block_num = 0;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_difficulty(netuid, start_diff);
         SubtensorModule::set_adjustment_interval(netuid, adjustment_interval);
@@ -545,7 +543,6 @@ fn test_burn_adjustment_case_d() {
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
 
         // Register key 1. This is a BURN registration
         let hotkey_account_id_1 = U256::from(1);
@@ -619,6 +616,7 @@ fn test_burn_adjustment_case_e() {
         let target_registrations_per_interval: u16 = 3;
         let start_diff: u64 = 10_000;
         let mut curr_block_num = 0;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_max_registrations_per_block(netuid, 10);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_difficulty(netuid, start_diff);
@@ -627,7 +625,6 @@ fn test_burn_adjustment_case_e() {
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
 
         // Register key 1. This is a POW registration
         let hotkey_account_id_1 = U256::from(1);
@@ -692,6 +689,7 @@ fn test_burn_adjustment_case_f() {
         let target_registrations_per_interval: u16 = 1;
         let start_diff: u64 = 10_000;
         let mut curr_block_num = 0;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_max_registrations_per_block(netuid, 10);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_difficulty(netuid, start_diff);
@@ -700,7 +698,6 @@ fn test_burn_adjustment_case_f() {
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
 
         // Register key 1. This is a POW registration
         let hotkey_account_id_1 = U256::from(1);
@@ -764,6 +761,7 @@ fn test_burn_adjustment_case_e_zero_registrations() {
         let adjustment_interval = 1;
         let target_registrations_per_interval: u16 = 1;
         let start_diff: u64 = 10_000;
+        add_network(netuid, tempo, 0);
         SubtensorModule::set_max_registrations_per_block(netuid, 10);
         SubtensorModule::set_burn(netuid, burn_cost);
         SubtensorModule::set_difficulty(netuid, start_diff);
@@ -772,7 +770,6 @@ fn test_burn_adjustment_case_e_zero_registrations() {
             netuid,
             target_registrations_per_interval,
         );
-        add_network(netuid, tempo, 0);
 
         // No registrations this interval of any kind.
         step_block(1);
@@ -792,286 +789,3 @@ fn test_burn_adjustment_case_e_zero_registrations() {
     });
 }
 
-#[test]
-fn test_emission_calculation_full() {
-    new_test_ext().execute_with(|| {
-        // Create network 1
-        let netuid1: u16 = 1;
-        let tempo1: u16 = 13;
-        add_network(netuid1, tempo1, 0);
-        for i in 0u16..5u16 {
-            let key = U256::from(i);
-            let stake = 100000;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid1, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid1, i, true);
-        }
-
-        step_block(1001);
-
-        assert_eq!(SubtensorModule::get_emission_value(netuid1), 1_000_000_000);
-    });
-}
-
-#[test]
-fn test_emission_calculation_half() {
-    new_test_ext().execute_with(|| {
-        // Create network 1
-        let netuid1: u16 = 1;
-        let tempo1: u16 = 13;
-        add_network(netuid1, tempo1, 0);
-        for i in 0u16..5u16 {
-            let key = U256::from(i);
-            let stake = 100000;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid1, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid1, i, true);
-        }
-
-        // Create network 2
-        let netuid2: u16 = 2;
-        let tempo2: u16 = 13;
-        add_network(netuid2, tempo2, 0);
-        for i in 6u16..11u16 {
-            let key = U256::from(i);
-            let stake = 100000;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid2, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid2, i, true);
-        }
-
-        step_block(1001);
-
-        assert_eq!(SubtensorModule::get_emission_value(netuid1), 500_000_000);
-        assert_eq!(SubtensorModule::get_emission_value(netuid2), 500_000_000);
-    });
-}
-
-#[test]
-fn test_emission_calculation_one_tenth() {
-    new_test_ext().execute_with(|| {
-        // Create network 1
-        let netuid1: u16 = 1;
-        let tempo1: u16 = 13;
-        add_network(netuid1, tempo1, 0);
-        for i in 0u16..5u16 {
-            let key = U256::from(i);
-            let stake = 1_000_000_000;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid1, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid1, i, true);
-        }
-
-        // Create network 2
-        let netuid2: u16 = 2;
-        let tempo2: u16 = 13;
-        add_network(netuid2, tempo2, 0);
-        for i in 6u16..11u16 {
-            let key = U256::from(i);
-            let stake = 111_111_111;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid2, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid2, i, true);
-        }
-
-        step_block(1001);
-
-        assert_eq!(SubtensorModule::get_emission_value(netuid1), 900000000);
-        assert_eq!(SubtensorModule::get_emission_value(netuid2), 100000000);
-    });
-}
-
-#[test]
-fn test_emission_calculation_with_empty_network() {
-    new_test_ext().execute_with(|| {
-        // Create network 1
-        let netuid1: u16 = 1;
-        let tempo1: u16 = 13;
-        add_network(netuid1, tempo1, 0);
-        for i in 0u16..5u16 {
-            let key = U256::from(i);
-            let stake = 1_000_000_000;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid1, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid1, i, true);
-        }
-
-        // Create network 2
-        let netuid2: u16 = 2;
-        let tempo2: u16 = 13;
-        add_network(netuid2, tempo2, 0);
-        for i in 6u16..11u16 {
-            let key = U256::from(i);
-            let stake = 111_111_111;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid2, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid2, i, true);
-        }
-
-        // Create network 3
-        let netuid3: u16 = 3;
-        let tempo3: u16 = 13;
-        add_network(netuid3, tempo3, 0);
-        for i in 12u16..17u16 {
-            let key = U256::from(i);
-            let stake = 0;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid3, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid3, i, true);
-        }
-
-        step_block(1001);
-
-        assert_eq!(SubtensorModule::get_emission_value(netuid1), 900000000);
-        assert_eq!(SubtensorModule::get_emission_value(netuid2), 100000000);
-        assert_eq!(SubtensorModule::get_emission_value(netuid3), 0);
-    });
-}
-
-#[test]
-fn test_emission_calculation_with_copied_stake() {
-    new_test_ext().execute_with(|| {
-        // Create network 1
-        let netuid1: u16 = 1;
-        let tempo1: u16 = 13;
-        add_network(netuid1, tempo1, 0);
-        for i in 0u16..5u16 {
-            let key = U256::from(i);
-            let stake = 1_000_000_000;
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid1, &key, 0);
-
-            SubtensorModule::increase_stake_on_coldkey_hotkey_account( &key, &key, stake );
-            SubtensorModule::set_validator_permit_for_uid(netuid1, i, true);
-        }
-
-        // Create network 2
-        let netuid2: u16 = 2;
-        let tempo2: u16 = 13;
-        add_network(netuid2, tempo2, 0);
-        for i in 0u16..5u16 {
-            let key = U256::from(i);
-
-            SubtensorModule::create_account_if_non_existent( &key, &key );
-            SubtensorModule::append_neuron(netuid2, &key, 0);
-
-            SubtensorModule::set_validator_permit_for_uid(netuid2, i, true);
-        }
-
-        step_block(1001);
-
-        assert_eq!(SubtensorModule::get_emission_value(netuid1), 500_000_000);
-        assert_eq!(SubtensorModule::get_emission_value(netuid2), 500_000_000);
-    });
-}
-
-// add_network( netuid1, tempo1, 0 );
-// add_network( netuid2, tempo2, 0 );
-
-// // // Lets step a block. There if no emission because we have not set an emission vector.
-// assert_eq!( SubtensorModule::get_pending_emission( netuid0 ), 0 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid1 ), 0 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid2 ), 0 );
-// step_block(1);
-// assert_eq!( SubtensorModule::get_pending_emission( netuid0 ), 0 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid1 ), 0 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid2 ), 0 );
-
-// // Lets set the block emission for this network. It will get all the emission.
-// let netuids: Vec<u16> = vec![ 0, 1, 2];
-// let emission: Vec<u64> = vec![ 333_333_333, 333_333_333, 333_333_334  ];
-// assert_ok!( SubtensorModule::sudo_set_emission_values(<<Test as Config>::RuntimeOrigin>::root(), netuids, emission) );
-
-// // Run a forward block. All emission ends up in pending.
-// assert_eq!( SubtensorModule::get_emission_value( netuid0 ), 333_333_333 );
-// assert_eq!( SubtensorModule::get_emission_value( netuid1 ), 333_333_333 );
-// assert_eq!( SubtensorModule::get_emission_value( netuid2 ), 333_333_334 );
-// step_block(1);
-// assert_eq!( SubtensorModule::get_pending_emission( netuid0 ), 333_333_333 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid1 ), 333_333_333 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid2 ), 333_333_334 );
-
-// // Run two more blocks and emission accrues for all networks.
-// step_block(1);
-// assert_eq!( SubtensorModule::get_pending_emission( netuid0 ), 666_666_666 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid1 ), 666_666_666 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid2 ), 666_666_668 );
-
-// step_block(1);
-// assert_eq!( SubtensorModule::get_pending_emission( netuid0 ), 999_999_999 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid1 ), 999_999_999 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid2 ), 1_000_000_002 );
-
-// // Create keys.
-// let hotkey0: u64 = 0;
-// let coldkey0: u64 = 0;
-
-// // Register 1 neuron to each network starting emission.
-// register_ok_neuron( netuid0, hotkey0, coldkey0, 39420842 );
-// register_ok_neuron( netuid1, hotkey0, coldkey0, 12412392 );
-// register_ok_neuron( netuid2, hotkey0, coldkey0, 21813123 );
-
-// // Run the block.
-// step_block(1);
-// assert_eq!( SubtensorModule::get_pending_emission( netuid0 ), 1_333_333_332 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid1 ), 1_333_333_332 );
-// assert_eq!( SubtensorModule::get_pending_emission( netuid2 ), 1_333_333_336 );
-
-// #[test]
-// fn test_nakamoto(){
-//     new_test_ext().execute_with(|| {
-
-//         // Create nakamoto.
-//         let n: u16 = 10;
-//         let netuid: u16 = 0;
-//         let tempo: u16 = 100;
-//         add_network( netuid, tempo, 0 );
-//         let netuids: Vec<u16> = vec![ 0 ];
-//         let emission: Vec<u64> = vec![ 1_000_000_000 ];
-//         assert_ok!( SubtensorModule::sudo_set_emission_values(<<Test as Config>::RuntimeOrigin>::root(), netuids, emission) );
-
-//         // Increase network size to 4096
-//         SubtensorModule::set_max_allowed_uids( netuid, n );
-//         SubtensorModule::set_max_registrations_per_block( netuid, n * 2 );
-
-//         // Register neurons.
-//         for i in 0..n as u64 {
-//             log::trace!( "Register:\n{:?}\n", i );
-//             register_ok_neuron( netuid, i, i, i * 1_000_000_000 + i * 1_000_000 );
-//             assert_eq!( SubtensorModule::get_hotkey_for_net_and_uid( netuid, i as u16 ).unwrap(), i );
-//         }
-
-//         // Register the next batch to replace the older ones.
-//         for i in 0..n as u64 {
-//             log::trace!( "Register:\n{:?}\n", i );
-//             register_ok_neuron( netuid, i + n as u64, i + n as u64 , i * 2_200_100_500 + i * 2_000_000 + 124124 );
-//             assert_eq!( SubtensorModule::get_hotkey_for_net_and_uid( netuid, i as u16 ).unwrap(), i + n as u64 );
-//         }
-
-//     });
-// }
