+++ pallets/subtensor/tests/difficulty.rs
@@ -14,11 +14,8 @@ fn test_registration_difficulty_adjustment() {
         assert_eq!(SubtensorModule::get_difficulty_as_u64(netuid), 10000); // Check initial difficulty.
         assert_eq!(SubtensorModule::get_last_adjustment_block(netuid), 0); // Last adjustment block starts at 0.
         assert_eq!(SubtensorModule::get_registrations_this_block(netuid), 0); // No registrations this block.
-        assert_eq!(
-            SubtensorModule::get_target_registrations_per_interval(netuid),
-            2
-        ); // Target is default.
-        assert_eq!(SubtensorModule::get_adjustment_interval(netuid), 100); // Default adustment intrerval.
+        SubtensorModule::set_target_registrations_per_interval(netuid, 2);
+        SubtensorModule::set_adjustment_interval(netuid,100);
         assert_eq!(
             SubtensorModule::get_network_registration_allowed(netuid),
             true
