+++ pallets/subtensor/tests/sudo.rs
@@ -7,50 +7,14 @@ use pallet_subtensor::{Error, Event};
 use sp_core::U256;
 
 #[test]
-fn test_defaults() {
-    new_test_ext().execute_with(|| {
-        let netuid = 0;
-        add_network(netuid, 10, 0);
-        assert_eq!(SubtensorModule::get_number_of_subnets(), 1); // There is a single network.
-        assert_eq!(SubtensorModule::get_subnetwork_n(netuid), 0); // Network size is zero.
-        assert_eq!(SubtensorModule::get_rho(netuid), 30);
-        assert_eq!(SubtensorModule::get_tempo(netuid), 10);
-        assert_eq!(SubtensorModule::get_kappa(netuid), 32_767);
-        assert_eq!(SubtensorModule::get_min_difficulty(netuid), 1);
-        assert_eq!(SubtensorModule::get_max_difficulty(netuid), u64::MAX);
-        assert_eq!(SubtensorModule::get_difficulty_as_u64(netuid), 10000);
-        assert_eq!(SubtensorModule::get_immunity_period(netuid), 2);
-        assert_eq!(SubtensorModule::get_emission_value(netuid), 0);
-        assert_eq!(SubtensorModule::get_activity_cutoff(netuid), 5000);
-        assert_eq!(SubtensorModule::get_pending_emission(netuid), 0);
-        assert_eq!(SubtensorModule::get_max_weight_limit(netuid), u16::MAX);
-        assert_eq!(SubtensorModule::get_max_allowed_uids(netuid), 2);
-        assert_eq!(SubtensorModule::get_min_allowed_weights(netuid), 0);
-        assert_eq!(SubtensorModule::get_adjustment_interval(netuid), 100);
-        assert_eq!(SubtensorModule::get_bonds_moving_average(netuid), 900_000);
-        assert_eq!(SubtensorModule::get_last_adjustment_block(netuid), 0);
-        assert_eq!(SubtensorModule::get_last_mechanism_step_block(netuid), 0);
-        assert_eq!(SubtensorModule::get_blocks_since_last_step(netuid), 0);
-        assert_eq!(SubtensorModule::get_registrations_this_block(netuid), 0);
-        assert_eq!(SubtensorModule::get_validator_prune_len(netuid), 0);
-        assert_eq!(SubtensorModule::get_scaling_law_power(netuid), 50);
-        assert_eq!(SubtensorModule::get_registrations_this_interval(netuid), 0);
-        assert_eq!(SubtensorModule::get_max_registrations_per_block(netuid), 3);
-        assert_eq!(
-            SubtensorModule::get_target_registrations_per_interval(netuid),
-            2
-        );
-    });
-}
-
-#[test]
 fn test_sudo_registration() {
     new_test_ext().execute_with(|| {
-        add_network(0, 0, 0);
+        let netuid: u16 = 1;
+        add_network(netuid, 0, 0);
         SubtensorModule::set_max_allowed_uids(0, 10);
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(0),
             U256::from(0),
             10,
@@ -58,7 +22,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(1),
             U256::from(1),
             10,
@@ -66,7 +30,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(2),
             U256::from(2),
             10,
@@ -74,7 +38,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(3),
             U256::from(3),
             10,
@@ -82,7 +46,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(4),
             U256::from(4),
             10,
@@ -90,7 +54,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(5),
             U256::from(5),
             10,
@@ -98,7 +62,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(6),
             U256::from(6),
             10,
@@ -106,7 +70,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(7),
             U256::from(7),
             10,
@@ -114,7 +78,7 @@ fn test_sudo_registration() {
         ));
         assert_ok!(SubtensorModule::sudo_register(
             <<Test as Config>::RuntimeOrigin>::root(),
-            0,
+            netuid,
             U256::from(8),
             U256::from(8),
             10,
@@ -130,39 +94,39 @@ fn test_sudo_registration() {
         assert_eq!(SubtensorModule::get_coldkey_balance(&U256::from(7)), 11);
         assert_eq!(SubtensorModule::get_coldkey_balance(&U256::from(8)), 11);
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 0).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 0).unwrap(),
             U256::from(0)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 1).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 1).unwrap(),
             U256::from(1)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 2).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 2).unwrap(),
             U256::from(2)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 3).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 3).unwrap(),
             U256::from(3)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 4).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 4).unwrap(),
             U256::from(4)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 5).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 5).unwrap(),
             U256::from(5)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 6).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 6).unwrap(),
             U256::from(6)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 7).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 7).unwrap(),
             U256::from(7)
         );
         assert_eq!(
-            SubtensorModule::get_hotkey_for_net_and_uid(0, 8).unwrap(),
+            SubtensorModule::get_hotkey_for_net_and_uid(netuid, 8).unwrap(),
             U256::from(8)
         );
         assert_eq!(SubtensorModule::get_total_stake(), 90);
@@ -231,8 +195,8 @@ fn test_sudo_set_min_difficulty() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_min_difficulty(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_min_difficulty(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_min_difficulty(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -264,8 +228,8 @@ fn test_sudo_set_max_difficulty() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_max_difficulty(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_max_difficulty(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_max_difficulty(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -297,8 +261,8 @@ fn test_sudo_set_weights_version_key() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_weights_version_key(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_weights_version_key(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_weights_version_key(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -330,8 +294,8 @@ fn test_sudo_set_weights_set_rate_limit() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_weights_set_rate_limit(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_weights_set_rate_limit(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_weights_set_rate_limit(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -369,8 +333,8 @@ fn test_sudo_set_adjustment_interval() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_adjustment_interval(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_adjustment_interval(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_adjustment_interval(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -402,8 +366,8 @@ fn test_sudo_set_adjustment_alpha() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_adjustment_alpha(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_adjustment_alpha(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_adjustment_alpha(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -430,14 +394,35 @@ fn test_sudo_set_adjustment_alpha() {
     });
 }
 
+
+#[test]
+fn test_sudo_subnet_owner_cut() {
+    new_test_ext().execute_with(|| {
+        let to_be_set: u16 = 10;
+        let init_value: u16 = SubtensorModule::get_subnet_owner_cut();
+        assert_eq!(
+            SubtensorModule::sudo_set_subnet_owner_cut(
+                <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
+                to_be_set
+            ),
+            Err(DispatchError::BadOrigin.into())
+        );
+        assert_eq!(SubtensorModule::get_subnet_owner_cut(), init_value);
+        assert_ok!(SubtensorModule::sudo_set_subnet_owner_cut(
+            <<Test as Config>::RuntimeOrigin>::root(),
+            to_be_set
+        ));
+        assert_eq!(SubtensorModule::get_subnet_owner_cut(), to_be_set);
+    });
+}
+
 #[test]
 fn test_sudo_validator_prune_len() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_validator_prune_len(netuid);
         add_network(netuid, 10, 0);
-
+        let init_value: u64 = SubtensorModule::get_validator_prune_len(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_validator_prune_len(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -469,8 +454,8 @@ fn test_sudo_set_scaling_law_power() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 50;
-        let init_value: u16 = SubtensorModule::get_scaling_law_power(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_scaling_law_power(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_scaling_law_power(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -502,8 +487,8 @@ fn test_sudo_set_max_weight_limit() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_max_weight_limit(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_max_weight_limit(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_max_weight_limit(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -554,8 +539,8 @@ fn test_sudo_set_immunity_period() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_immunity_period(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_immunity_period(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_immunity_period(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -587,8 +572,8 @@ fn test_sudo_set_min_allowed_weights() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_min_allowed_weights(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_min_allowed_weights(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_min_allowed_weights(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -620,8 +605,8 @@ fn test_sudo_set_max_allowed_uids() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_max_allowed_uids(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_max_allowed_uids(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_max_allowed_uids(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -653,8 +638,8 @@ fn test_sudo_set_and_decrease_max_allowed_uids() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_max_allowed_uids(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_max_allowed_uids(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_max_allowed_uids(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -690,8 +675,8 @@ fn test_sudo_set_kappa() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_kappa(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_kappa(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_kappa(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -723,8 +708,8 @@ fn test_sudo_set_rho() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_rho(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_rho(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_rho(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -756,8 +741,8 @@ fn test_sudo_set_activity_cutoff() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_activity_cutoff(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_activity_cutoff(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_activity_cutoff(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -789,8 +774,8 @@ fn test_sudo_set_target_registrations_per_interval() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_target_registrations_per_interval(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_target_registrations_per_interval(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_target_registrations_per_interval(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -828,8 +813,8 @@ fn test_sudo_set_difficulty() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_difficulty_as_u64(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_difficulty_as_u64(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_difficulty(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -861,8 +846,8 @@ fn test_sudo_set_max_allowed_validators() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u16 = 10;
-        let init_value: u16 = SubtensorModule::get_max_allowed_validators(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u16 = SubtensorModule::get_max_allowed_validators(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_max_allowed_validators(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -900,8 +885,8 @@ fn test_sudo_set_bonds_moving_average() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_bonds_moving_average(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_bonds_moving_average(netuid);
         assert_eq!(
             SubtensorModule::sudo_set_bonds_moving_average(
                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
@@ -931,103 +916,103 @@ fn test_sudo_set_bonds_moving_average() {
     });
 }
 
-#[test]
-fn test_sudo_set_network_connection_requirement() {
-    new_test_ext().execute_with(|| {
-        let netuid_a: u16 = 1;
-        let netuid_b: u16 = 2;
-        let requirement: u16 = u16::MAX;
-        assert_eq!(
-            SubtensorModule::sudo_add_network_connection_requirement(
-                <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
-                netuid_a,
-                netuid_b,
-                requirement
-            ),
-            Err(DispatchError::BadOrigin.into())
-        );
-        assert_eq!(
-            SubtensorModule::sudo_add_network_connection_requirement(
-                <<Test as Config>::RuntimeOrigin>::root(),
-                netuid_a,
-                netuid_b,
-                requirement
-            ),
-            Err(Error::<Test>::NetworkDoesNotExist.into())
-        );
-        add_network(netuid_a, 10, 0);
-        assert_eq!(
-            SubtensorModule::sudo_add_network_connection_requirement(
-                <<Test as Config>::RuntimeOrigin>::root(),
-                netuid_a,
-                netuid_a,
-                requirement
-            ),
-            Err(Error::<Test>::InvalidConnectionRequirement.into())
-        );
-        assert_eq!(
-            SubtensorModule::sudo_add_network_connection_requirement(
-                <<Test as Config>::RuntimeOrigin>::root(),
-                netuid_a,
-                netuid_b,
-                requirement
-            ),
-            Err(Error::<Test>::NetworkDoesNotExist.into())
-        );
-        add_network(netuid_b, 10, 0);
-        assert_ok!(SubtensorModule::sudo_add_network_connection_requirement(
-            <<Test as Config>::RuntimeOrigin>::root(),
-            netuid_a,
-            netuid_b,
-            requirement
-        ));
-        assert_eq!(
-            SubtensorModule::get_network_connection_requirement(netuid_a, netuid_b),
-            requirement
-        );
-        assert_eq!(
-            SubtensorModule::sudo_remove_network_connection_requirement(
-                <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
-                netuid_a,
-                netuid_b
-            ),
-            Err(DispatchError::BadOrigin.into())
-        );
-        assert_eq!(
-            SubtensorModule::sudo_remove_network_connection_requirement(
-                <<Test as Config>::RuntimeOrigin>::root(),
-                5 as u16,
-                5 as u16
-            ),
-            Err(Error::<Test>::NetworkDoesNotExist.into())
-        );
-        assert_eq!(
-            SubtensorModule::sudo_remove_network_connection_requirement(
-                <<Test as Config>::RuntimeOrigin>::root(),
-                netuid_a,
-                5 as u16
-            ),
-            Err(Error::<Test>::NetworkDoesNotExist.into())
-        );
-        assert_ok!(SubtensorModule::sudo_remove_network_connection_requirement(
-            <<Test as Config>::RuntimeOrigin>::root(),
-            netuid_a,
-            netuid_b
-        ));
-        assert_eq!(
-            SubtensorModule::network_connection_requirement_exists(netuid_a, netuid_b),
-            false
-        );
-    });
-}
+// DEPRECATED #[test]
+// fn test_sudo_set_network_connection_requirement() {
+//     new_test_ext().execute_with(|| {
+//         let netuid_a: u16 = 1;
+//         let netuid_b: u16 = 2;
+//         let requirement: u16 = u16::MAX;
+//         assert_eq!(
+//             SubtensorModule::sudo_add_network_connection_requirement(
+//                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
+//                 netuid_a,
+//                 netuid_b,
+//                 requirement
+//             ),
+//             Err(DispatchError::BadOrigin.into())
+//         );
+//         assert_eq!(
+//             SubtensorModule::sudo_add_network_connection_requirement(
+//                 <<Test as Config>::RuntimeOrigin>::root(),
+//                 netuid_a,
+//                 netuid_b,
+//                 requirement
+//             ),
+//             Err(Error::<Test>::NetworkDoesNotExist.into())
+//         );
+//         add_network(netuid_a, 10, 0);
+//         assert_eq!(
+//             SubtensorModule::sudo_add_network_connection_requirement(
+//                 <<Test as Config>::RuntimeOrigin>::root(),
+//                 netuid_a,
+//                 netuid_a,
+//                 requirement
+//             ),
+//             Err(Error::<Test>::InvalidConnectionRequirement.into())
+//         );
+//         assert_eq!(
+//             SubtensorModule::sudo_add_network_connection_requirement(
+//                 <<Test as Config>::RuntimeOrigin>::root(),
+//                 netuid_a,
+//                 netuid_b,
+//                 requirement
+//             ),
+//             Err(Error::<Test>::NetworkDoesNotExist.into())
+//         );
+//         add_network(netuid_b, 10, 0);
+//         assert_ok!(SubtensorModule::sudo_add_network_connection_requirement(
+//             <<Test as Config>::RuntimeOrigin>::root(),
+//             netuid_a,
+//             netuid_b,
+//             requirement
+//         ));
+//         assert_eq!(
+//             SubtensorModule::get_network_connection_requirement(netuid_a, netuid_b),
+//             requirement
+//         );
+//         assert_eq!(
+//             SubtensorModule::sudo_remove_network_connection_requirement(
+//                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(0)),
+//                 netuid_a,
+//                 netuid_b
+//             ),
+//             Err(DispatchError::BadOrigin.into())
+//         );
+//         assert_eq!(
+//             SubtensorModule::sudo_remove_network_connection_requirement(
+//                 <<Test as Config>::RuntimeOrigin>::root(),
+//                 5 as u16,
+//                 5 as u16
+//             ),
+//             Err(Error::<Test>::NetworkDoesNotExist.into())
+//         );
+//         assert_eq!(
+//             SubtensorModule::sudo_remove_network_connection_requirement(
+//                 <<Test as Config>::RuntimeOrigin>::root(),
+//                 netuid_a,
+//                 5 as u16
+//             ),
+//             Err(Error::<Test>::NetworkDoesNotExist.into())
+//         );
+//         assert_ok!(SubtensorModule::sudo_remove_network_connection_requirement(
+//             <<Test as Config>::RuntimeOrigin>::root(),
+//             netuid_a,
+//             netuid_b
+//         ));
+//         assert_eq!(
+//             SubtensorModule::network_connection_requirement_exists(netuid_a, netuid_b),
+//             false
+//         );
+//     });
+// }
 
 #[test]
 fn test_sudo_set_rao_recycled() {
     new_test_ext().execute_with(|| {
         let netuid: u16 = 1;
         let to_be_set: u64 = 10;
-        let init_value: u64 = SubtensorModule::get_rao_recycled(netuid);
         add_network(netuid, 10, 0);
+        let init_value: u64 = SubtensorModule::get_rao_recycled(netuid);
 
         // Need to run from genesis block
         run_to_block(1);
@@ -1112,11 +1097,7 @@ fn test_sudo_test_tempo_pending_emissions_ok() {
         assert_eq!(SubtensorModule::get_emission_value(netuid3), 0);
         let netuids: Vec<u16> = vec![1, 2, 3, 5];
         let emission: Vec<u64> = vec![100000000, 400000000, 200000000, 300000000];
-        assert_ok!(SubtensorModule::sudo_set_emission_values(
-            <<Test as Config>::RuntimeOrigin>::root(),
-            netuids,
-            emission
-        ));
+        SubtensorModule::set_emission_values( &netuids, emission.clone() );
         assert_eq!(SubtensorModule::get_emission_value(netuid0), 100000000);
         assert_eq!(SubtensorModule::get_emission_value(netuid1), 400000000);
         assert_eq!(SubtensorModule::get_emission_value(netuid2), 200000000);
@@ -1143,16 +1124,7 @@ pub fn test_sudo_test_pending_emission_ok() {
         add_network(netuid1, tempo1, 0);
         add_network(netuid2, tempo2, 0);
 
-        assert_ok!(SubtensorModule::sudo_set_emission_values(
-            <<Test as Config>::RuntimeOrigin>::root(),
-            netuids,
-            emission
-        ));
+        SubtensorModule::set_emission_values( &netuids, emission.clone() );
         assert_eq!(SubtensorModule::get_emission_value(netuid1), 250000000);
-
-        step_block(3);
-
-        assert_eq!(SubtensorModule::get_pending_emission(netuid1), 0); // emission drained at block 3 for tempo 5
-        assert_eq!(SubtensorModule::get_pending_emission(netuid2), 2250000000); // 750000000 + 750000000 + 750000000
     });
 }
